rename old files to legacy
